requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
yfinance==0.2.28
pydantic==1.8.2
python-dotenv==0.19.0
requests==2.26.0
whisper==1.0.0
gTTS==2.3.1
python-multipart==0.0.5
aiofiles==0.8.0
faiss-cpu==1.7.4
numpy>=1.19.5
langchain==0.0.184
openai==0.27.0
httpx==0.24.0
streamlit==1.24.0